[Sapling][Bug] Fixing multi-source notes spending
Fixing the currently not working shielded spending from multiple notes with different spending keys. Plus, added test coverage for this scenario.
Essentially, the issue was that we were using the same spending key for all of the spending proofs. It was a remanent from the previous workflow when we only were able to spend shielded coins from a single source.

This obviously enables the new notes coin control feature that is being introduced in 1963. So.. as soon as we merge this one, will rebase 1963 on top.

Extra note: the functional test coverage was added before the fix so it can easily be checked before and after the solution.