[Core] Masternode collateral index
Currently we check whether a masternode collateral is spent, in CMasternode::Check(), which is called in a time-based polling fashion, looking for the collateral outpoint in the coins view cache (and countless other times on-demand, when we want an up-to-date status).
This is highly expensive, and inconsistent (peers check masternodes at different times, thus have slightly different masternode states).

Collaterals, as all coins/outpoints, can only be spent inside blocks.

So we should do this check only when connecting new blocks, not every x seconds, or every time that we need to count the number of enabled masternodes, just to name the most glaring example (CMasternodeMan::CountEnabled).
This can be done very efficiently if we change the data-structure to store them inside the manager, from a vector, to a map, indexed by collateral outpoints.

Second part, to be addressed in a follow-up PR:

After this change, CMasternode::Check() includes only the ping-checks, which are so fast that can be done in "real-time", thus completely removing the need to cache the activeState inside the CMasternode class, and to call Check() every time that we want to update it