[BUG] g_IsSapling active flag during reindex
Problem:
Starting the wallet with --reindex (w/wo --zapwallettxes) causes a crash if the wallet has shield transactions.

Cause:
g_IsSaplingActive is used to select the proper serialization for v2 txes (with or without sapData).
During init, g_IsSaplingActive is set to true in LoadBlockIndexDB when connecting a post-enforcement block.
When --reindexing, though, LoadBlockIndexDB is skipped in LoadBlockIndex, and the global flag remains false.
Therefore g_IsSaplingActive is false when init gets to InitLoadWallet and the deserialization of shield transactions throws.

Solution:
Set g_IsSaplingActive to true when --reindex.
This way the wallet can be loaded, and then the flag switches back to false as soon as the first pre-enforcement block is connected (ThreadImport starts only after the wallet has been loaded).

Note: A wallet would now crash during a reindex, if containing a pre-enforcement v2 transaction (old serialization, no sapData).
But such transactions are non-standard (afaik there's only one on the whole main-net blockchain).
A wallet with this transaction (if it still exists) would not be able to use --reindex and would be forced to --resync.