[BUG][GUI] Encrypted memo with transparent recipient address
In the send widget, we are presenting a warning if a t->t tx includes any memo.
But this is not the correct logic, as it doesn't cover all possible mis-placements:

    a t->s tx could still have the memo set on a transparent recipient. The warning is not triggered though, if the transaction has at least one shielded recipient (thus it's a shielded tx).
    a s->t tx should not be able to include any memo, yet no warning is triggered.

Memos are encrypted in the ciphertext of the output description, so the only thing that matters here are the outputs.
We should check the presence of the memo, for each output, against the recipient address of that particular output.