[RPC] redirect sendmany to shieldedsendmany when recipient is shielded
Instead of throwing an error, let's repack the json request and send it to shieldedsendmany.
This makes the integration of send-to-shield-address much easier for third-party services.

We keep the old implementation of sendmany for t->t transactions, until the two flows are properly abstracted.

Also fix a minor bug found in sendmany (not properly considering the fIncludeDelegated flag), and add capability to spend P2CS inside SaplingOperation.

Note: the same thing proposed here, could be done with the RPC sendtoaddress (future).