[BUG] Fix lock order issue with CMMan::CheckAndRemove and CMPayments::CleanPaymentList
Remove the "cleanup" during DB read (move it at the beginning of the thread).
Pass the masternode-count and current-height to CleanPaymentList, so there is no call to the masternode manager from within it.

Note: GA will keep failing as there are other locking issues. They will be addressed in separate PR(s).