[Sapling] Try to recover corruption of notes cache during send operation
This is a contingency plan, in case of missing cached nullifier of a spent note in the wallet.
Since, during the spend, the wallet is unlocked, and we have the spending key, we can directly recover the nullifier from the note (provided that we have a valid cached witness) and fix the issue in the background, without requiring any user intervention.

If we don't have the witness, we can still provide a better log and error message in the gui, than a generic, and probably obscure

bad-txns-shielded-requirements-not-met