[Build] Package sapling params in repo/release
Since we no longer require the sprout params, we can package the sapling params inside the source repository (and the release binaries) until such time that the client has the ability to fetch these params directly from the P2P network.

Overview of the changes here per-OS:

    macOS release binaries will use the params that are bundled into the .app bundle
    Windows installer .exe will automatically install the params into the AppData directory.
    Linux PPA/Snap/Copr installs will automatically install the params into the appropriate directory (either $SNAP_USER_COMMON/.pivx-params or /usr/local/share)
    Linux tarballs include a bash shell script that can be run to install the params files into $HOME/.pivx-params

Note: Windows users that choose to use the .zip binary package will need to manually copy the params files to the appropriate AppData directory on their own (to be documented later)

The install-params.sh script is also run-able on linux/macOS from repository clones or source tarball paths to install the params into the OS-appropriate location for future use.