Update mnmodel.cpp
As # 2019 got little bit big and we are entering into a freeze period, this is a quick fix to the "MISSING" state invalidly shown as "REMOVE" in the masternodes screen. Introduced by # 2006.
This affects the GUI as it makes masternodes that are not started to appear as "REMOVE" instead of "MISSING" as it was before (could be a MN in the configuration file only or even with the collateral accepted but not started).