[Wallet] BugFix: automatic upgrade for unlocked wallet
Essentially there is a problem in unlocked wallets running HD version when them start running v5. The SSPKM seed is set but the version isn't being updated to reflect the sapling features support. Ending up with a wallet with all of the sapling capabilities with the only exception of marking spent notes as spent..

This fixes it plus adds a guard into the activation mechanism to be a 100% sure that the version is set when the SSPKM is activated.