[Masternode] Encapsulate global and do not parse privkey redundantly everywhere.
Performance improvement for masternodes, instead of having to parse the global private key string, decode it once and cache it on the active masternode class, plus remove the global mn privkey string.
As this was decoded and re-decoded everywhere, this will have a direct effect on the ping send process, vote on finalized budgets, masternode payments block processing, and budget proposal vote flow.