[BUG][MN] Immediately detect forked masternodes
Check the pinged blockhash at the beginning of CMasternodePing::CheckAndUpdate.
So we don't ask new broadcasts when it's wrong, and we can immediately detect stale mnbs.

Also, explicitly pass the signature time when constructing CMasternode and CMasternodePing objects (instead of automatically setting it to the current time, which might be error prone, leading to unintended updates).