[Consensus] Max block space available for shielded txes
Given the results of the stress tests performed, limit it to 750 kB (at least for this first iteration).
Add functional test.

Note: this causes a fork on testnet4 if an "old" client mines a block with more than 750kB (cumulative size) of shielded txes.
Although, in order to do so, the staker should set a -blockmaxsize at startup, higher than the default (and there should be enough txes available in the mempool).