[Script][BUG] Fix signature malleability for t inputs in Sapling txes
We are currently signing the transaparent inputs with SIGVERSION_BASE, thus we are not considering the sapData in the hash.

The functional test included (sapling_malleable_sigs.py) provides a demo for a basic trolling exploit (intercept a t->z tx to effectively burn the amount of PIV that was supposed to be shielded).

Fix:
Always use SIGVERSION_SAPLING to sign txes with version >= TxVersion::SAPLING.
This required a bit of refactoring, in order to pass the proper sigversion.
In the future we could simplify this flow, saving the sigversion directly inside the SignatureCreator.

The json data for sighash_test unit test has been updated to reflect the new hashes.

The sapling_malleable_sigs functional test, now failing after the fix, is updated to verify that the proper exception is thrown.