[BUG][Masternode] Miner not paying valid finalized budget if <20 active masternodes
Edge case where the consensus is not aligned between the block producer and the block verifier.

Verifier: (CheckBlock --> IsBlockPayeeValid --> IsBudgetPaymentBlock), during a superblock, expects a payment to a budget proposal, even with the case of N<20 active masternodes on the network, provided that all N voted on the finalized budget (the threshold is N-1).

The block producer, instead, is paying a masternode (GetPayeeAndAmount).
Fix it, using the same flow of the verifier: first check the threshold, adjusting it for the case of N<20 (5% = 0), setting it to N-1, and then check that there is a finalized budget with enough votes.

Also remove the extra call to IsBudgetPaymentBlock directly from FillBlockPayee (now called by CBudgetManager::FillBlockPayee --> GetPayeeAndAmount).