[GUI] Settings information, fix missing initial masternodes count value.
The count isn't updated on every widget show, only via the event (which only happen every 40 minutes as Masternodes updates aren't regular at all).
This PR fixes it updating the masternode count every time that the screen is visible, then the event is in charge of update it if the user stays there long enough.