[RPC] Use "shield" nomenclature instead of "shielded"
Alternative to # 2077 that limits the scope to the RPC interface (and unit/functional tests that exercise the RPC interface)

Drop the "ed" suffix in the RPC scope, changes commands like
getnewshieldedaddress to getnewshieldaddress.

Also in blockchain informational calls; shielded_pool_value and
shieldedsupply become shield_pool_value and shieldsupply
respectively.