[GUI][BugFix] Duplicate transaction records
Finally found it, fixing the duplicate record issue.

Briefly explained: As is now, decomposeDebitTransaction can return false even when the outputs parsing was done correctly due the non-existency of shielded outputs. If this happens, the caller decomposeTransaction process it as if decomposeDebitTransaction didnt parse it successfully, appending an extra all-in-one record.