[BUG][GUI] Don't return StakingOnlyUnlocked from WalletModel::sendCoins
Bug found in current master.
Steps to reproduce

    Unlock the wallet for staking only.
    Try to send a transaction
    The wallet asks to insert the password (to fully unlock).
    Insert the password, and then accept the confirmation dialog.

Issue:
Instead of sending the tx, the wallet asks again for the password, and still doesn't send the transaction (without any feedback) even when the password is entered a second time.

Cause:
By the time the flow reaches WalletModel::sendCoins, the UnlockContext is lost, and the wallet has been relocked (for staking only) already.
This is correct, as the transaction is created and signed in WalletModel::prepareTransaction, and sendCoins only commits the transaction, therefore it doesn't need a fully-unlocked wallet.

The function, though, currently returns early with return value SendCoinsReturn::StakingOnlyUnlocked, if the wallet is unlocked for-staking-only, and the transaction isn't sent.

Fix:
Simply remove the un-needed check in sendCoins.