[Consensus] Max payment count for budget proposals
Set it to 6 on main-net and 20 on testnet.
Consider "expired" a proposal 1 and a half cycles after the last payment (instead of 1 cycle + a number of blocks equal to the number of payments) and enforce that block start is a superblock (currently checked only in the RPC, but not in consensus).
Update the functional test to check improper inputs during the creation of a budget proposal.

Also fix a bug with invalid total budget amount on testnet.