[GUI] Support for op_return proposal output records.
As the GUI is showing OP_RETURN output records as "no information", proposals fee outputs aren't being presented properly.

Proposal fee outputs only store the proposal hash in the op_return data and can't be distinguished only with the network data (proposals data is ephemeral), this PR aims to tackle it storing the proposal name inside the "comments" field of the wallet transaction.