Update settingsconsolewidget.cpp
Fix an issue with the double confirmation dialog introduced in # 1928

Currently the confirmation pop-up is shown only when the RPC command string is exactly "dumpwallet"/"dumpprivkey" (with no arguments).
Such command strings are not valid, as both RPC functions require one argument (the filename for "dumpwallet" and the PIVX address for "dumpprivkey").

So, when the user confirms the dialog, the RPC console returns an error (as the argument is missing), showing the RPC help.

When, instead, the user enters the correct RPC strings (dumpwallet <filename>/dumpprivkey <pivxaddress>), then the warning dialog is NOT presented and the command is executed immediately.

Fix it by doing a pre-validation for the potentially dangerous commands.
Bonus: include exportsaplingkey to the list.

Note: for dumpprivkey/exportsaplingkey this means a double parsing of the command string and decoding of the address argument (one during the pre-validation, and one during the actual execution of the RPC). But I think we can live with that for now.