[RPC] Fix inconsistencies in outputs
This is a subset of a larger effort to make sure all of our in-wallet RPC documentation and output is consistent with the JSON standard. In this particular PR, I've focused on dead code removal, and missing documentation.

Commits have been kept as atomic as possible to ease review, though the entire diff can be reviewed as well.