Update wallet_zkeys_tests.cpp
Follow up investigation to # 2116. Based on GA results, the diversified address is equal to the default address.
The code now will try to create a different address to the default one in a loop increasing the diversifier by one on each round.

This issue will need a deeper investigation in the future if/when we add diversified addresses features (at this moment, is pretty irrelevant as we aren't using diversified addresses at all). I have seen it only happening randomly on GA on linux and doesn't seems to be a library issue, if we provide the same seed to the process, the error isn't being thrown. It seems more to be an issue with the data pointer.