Generic event handler interface
A set of commits decoupled from # 1973, solving # 1840.

The final goal is # 1973 but the scope there got pretty big and is requiring more previous back ports (like 9725 among others that are an entire branch of back ports by themselves) so.. better to start merging this sooner rather than later to not continue having issues on macOS deployment (or well, more specifically issues with newer boost versions).