[GUI] send widget, hide contacts menu when clear event is triggered.
Inside the send screen the "clear all" event removes every row and there by every entry from the widget, as the contacts menu stores a local pointer to the entry where is being open and the data being pointed does't exist anymore, it was occasioning a crash there.
The solution is simple, close the menu contacts when the clear button is pressed.