[GUI] Use wchar_t API explicitly on Windows
backport of bitcoin # 13734 to use the wchar API for windows startup function. This is the first in a long list of upstream PRs that aim to bring full Unicode support to Windows clients as detailed in bitcoin # 13869

Also included here is the relevant parts of bitcoin # 5793, to use per-network auto startup shortcuts.